Update RunnerProfitAndLost.cs
IfWin and IfLose should be double not long.

see here https://api.developer.betfair.com/services/webapps/docs/display/1smk3cen4v3lu3yomq5qye0ni/Betting+Type+Definitions#BettingTypeDefinitions-RunnerProfitAndLoss

+++ cSharp/Api-ng-sample-code/Api-ng-sample-code/TO/RunnerProfitAndLost.cs
@@ -12,10 +12,10 @@ namespace Api_ng_sample_code.TO
         public long SelectionId { get; set; }
 
         [JsonProperty(PropertyName = "ifWin")]
-        public long IfWin { get; set; }
+        public double IfWin { get; set; }
 
         [JsonProperty(PropertyName = "ifLose")]
-        public long IfLose { get; set; }
+        public double IfLose { get; set; }
 
         public override string ToString()
         {
